rename - #124 : PostList 및 SeriesList 파일 이름 PostCard, SeriesCard로 변경
